Describe the learning modes that actually exist
The bot description and /help still advertised multiple choice and
listening, which stay hidden until they are finished. List the modes the
menu really offers instead.

While here: drop the dead noqa directives, narrow the difficulty cast to
the errors it can actually raise, and sort two import blocks.

+++ chinese_telegram_bot/app/utils/text.py
@@ -6,7 +6,7 @@ from datetime import datetime, timezone
 def progress_bar(value: float, total: float, width: int = 12) -> str:
     if total <= 0:
         return "▱" * width
-    filled = max(0, min(width, int(round(width * value / total))))
+    filled = max(0, min(width, round(width * value / total)))
     return "▰" * filled + "▱" * (width - filled)
